submission/requirements.txt
python>=3.11.7
accelerate==1.2.1
jiwer==3.0.5
numpy==1.24.4
opencv-python==4.9.0.80
pandas==1.5.3
safetensors==0.4.5
sympy==1.13.1
ultralytics==8.3.158
torch==2.6.0+cu124
torchvision==0.21.0+cu124
transformers==4.51.3
